refactor

+++ src/main/java/com/nft/app/entity/WithdrawRequest.java
@@ -18,8 +18,8 @@ public class WithdrawRequest {
   private String id;
   private String email;
   private Integer amount;
-  private Integer serviceCharge;
-  private Integer totalAmount;
+  private Double serviceCharge;
+  private Double totalAmount;
   private String status;
   private String comment;
 
@@ -32,7 +32,7 @@ public class WithdrawRequest {
   public WithdrawRequest(String email, Integer amount) {
     this.email = email;
     this.amount = amount;
-    this.serviceCharge = (int) Math.ceil(0.05 * amount);
+    this.serviceCharge = Math.ceil(0.05 * amount);
     this.totalAmount = amount + serviceCharge;
     this.status = "PENDING";
   }
